Monitor/requirements.txt
waybackpy
openpyxl
requests
pandas
beautifulsoup4
pyttsx3
python-dotenv
logging
boto3
slack_sdk
xlsxwriter
